+++ apps/tilbakemelding-api/src/main/java/no/nav/registre/testnorge/tilbakemeldingapi/TilbakemeldingApiApplicationStarter.java
@@ -7,7 +7,6 @@ import org.springframework.context.annotation.Import;
 import no.nav.registre.testnorge.libs.autodependencyanalysis.config.AutoRegistrationDependencyAnalysisConfiguration;
 
 @SpringBootApplication
-@Import(AutoRegistrationDependencyAnalysisConfiguration.class)
 public class TilbakemeldingApiApplicationStarter {
     public static void main(String[] args) {
         SpringApplication.run(TilbakemeldingApiApplicationStarter.class, args);
