+++ apps/tilbakemelding-api/src/main/java/no/nav/registre/testnorge/tilbakemeldingapi/config/AppConfig.java
@@ -1,11 +1,15 @@
 package no.nav.registre.testnorge.tilbakemeldingapi.config;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
+import javax.annotation.PostConstruct;
+
 import no.nav.registre.testnorge.libs.autodependencyanalysis.config.AutoRegistrationDependencyAnalysisConfiguration;
+import no.nav.registre.testnorge.libs.autodependencyanalysis.service.ApplicationRegistrationService;
 import no.nav.registre.testnorge.libs.core.config.ApplicationCoreConfig;
 import no.nav.registre.testnorge.libs.core.config.ApplicationProperties;
 import no.nav.registre.testnorge.libs.oauth2.config.SecureOAuth2ServerToServerConfiguration;
@@ -14,9 +18,12 @@ import no.nav.registre.testnorge.libs.slack.consumer.SlackConsumer;
 @Configuration
 @Import({
         ApplicationCoreConfig.class,
-        SecureOAuth2ServerToServerConfiguration.class
+        SecureOAuth2ServerToServerConfiguration.class,
+        AutoRegistrationDependencyAnalysisConfiguration.class
 })
+@RequiredArgsConstructor
 public class AppConfig {
+    private final ApplicationRegistrationService applicationRegistrationService;
 
     @Bean
     public SlackConsumer slackConsumer(
@@ -26,5 +33,10 @@ public class AppConfig {
             ApplicationProperties properties
     ) {
         return new SlackConsumer(token, baseUrl, proxyHost, properties.getName());
+    }
+
+    @PostConstruct
+    public void init() {
+        applicationRegistrationService.register();
     }
 }
